migrate to a "modern" Rock layout
