--- canbus.pc.in
@@ -1,12 +1,0 @@
-prefix=@CMAKE_INSTALL_PREFIX@
-exec_prefix=@CMAKE_INSTALL_PREFIX@
-libdir=${prefix}/lib
-includedir=${prefix}/include
-
-Name: canbus
-Description: Driver to access the CAN bus
-Version: @PROJECT_VERSION@
-Requires: base-types
-Libs: -L${libdir} -lcanbus
-Cflags: -I${includedir}
-
