+++ src/Message.hpp
@@ -4,24 +4,24 @@
 #include <base/Time.hpp>
 #include <stdint.h>
 
-#define CAN_ERR_OK  		0x00
-#define CAN_ERR_XMTFULL  	0x01
-#define CAN_ERR_OVERRUN  	0x02
-#define CAN_ERR_BUSERR 		0x04
-#define CAN_ERR_BUSOFF  	0x08
-#define CAN_ERR_RX_OVERFLOW	0x10
-#define CAN_ERR_TX_OVERFLOW 	0x20
+#define CAN_ERR_OK              0x00
+#define CAN_ERR_XMTFULL         0x01
+#define CAN_ERR_OVERRUN         0x02
+#define CAN_ERR_BUSERR          0x04
+#define CAN_ERR_BUSOFF          0x08
+#define CAN_ERR_RX_OVERFLOW     0x10
+#define CAN_ERR_TX_OVERFLOW     0x20
 
 namespace canbus
 {
     enum DRIVER_TYPE
     {
-	SOCKET,
-	HICO,
-	HICO_PCI,
-	VS_CAN,
-	CAN2WEB,
-	NET_GATEWAY
+        SOCKET,
+        HICO,
+        HICO_PCI,
+        VS_CAN,
+        CAN2WEB,
+        NET_GATEWAY
     };
     
     struct Message
