+++ src/backwards/canbus.hh
@@ -1,0 +1,7 @@
+#ifndef CANBUS_BACKWARDS_CANBUS_HH
+#define CANBUS_BACKWARDS_CANBUS_HH
+
+#include <canbus/Driver.hpp>
+
+#endif
+
