+++ src/backwards/canmessage.hh
@@ -1,0 +1,7 @@
+#ifndef CANBUS_BACKWARDS_MESSAGE_HH
+#define CANBUS_BACKWARDS_MESSAGE_HH
+
+#include <canbus/Message.hpp>
+
+#endif
+
